// frontend/src/types/Supplier.ts
// export interface Supplier {
//   id: number;
//   name: string;
//   email: string;
//   phone: string;
//   city: string;
// }
export interface Supplier {
  id: number;

  name: string;

  description: string;

  address: string;

  phone: string;

  email: string;

  licenseNo: string;

  licenseValidTill: string;
}